GROW-21 lint files
