+++ src/constants/index.ts
@@ -1,6 +1,8 @@
-import Layout from './layout';
-import Typography from './typography';
 import Colors from './colors';
+import Layout from './layout';
 import Spacing from './spacing';
+import Typography from './typography';
 
-export { Colors, Layout, Spacing, Typography };
+export {
+  Colors, Layout, Spacing, Typography,
+};
