+++ src/constants/layout.ts
@@ -5,5 +5,5 @@ export default StyleSheet.create({
     flex: 1,
     alignItems: 'center',
     justifyContent: 'center',
-  }
+  },
 });
